Modify login page and add new InputText component
