+++ src/app/components/LoginPage/PopupFailed.js
@@ -6,30 +6,21 @@ import PropTypes from 'prop-types';
 import CloseButton from '../../../assets/images/close-button.png';
 
 class PopupFailed extends Component {
-  constructor() {
-    super();
-    this.closePopup = this.closePopup.bind(this);
-  }
-
-  closePopup() {
-    const { closePopup } = this.props;
-    closePopup('');
-  }
-
   render() {
+    const { closePopup } = this.props;
     return (
       <div className={styles.popupBackground}>
         <div className={styles.failedPopup}>
           <div className={styles.titleBar}>
             <h5 className={styles.titleContent}>Đăng nhập không thành công</h5>
-            <Link to="/login" onClick={this.closePopup}>
+            <Link to="/login" onClick={closePopup}>
               <img className={styles.closeImg} src={CloseButton} alt="close" />
             </Link>
           </div>
           <div className={styles.content}>
             <p>Tên đăng nhập hoặc mật khẩu không chính xác. Vui lòng thử lại.</p>
           </div>
-          <Link className={styles.linkDecoration} to="/login" onClick={this.closePopup}>
+          <Link className={styles.linkDecoration} to="/login" onClick={closePopup}>
             <button className={styles.closeBtn} type="submit">
               OK
             </button>
