+++ src/app/components/LoginPage/index.js
@@ -3,50 +3,7 @@ import { Redirect } from 'react-router-dom';
 import PopupFailed from './PopupFailed';
 import styles from './login.scss';
 import Header from '../NavBar/Header';
-import PropTypes from 'prop-types';
-
-function InputText(props) {
-  const { label, divStyle, labelStyle, inputStyle, type, id, name, value, onChange } = props;
-  return (
-    <div className={divStyle}>
-      <label className={labelStyle} htmlFor={id}>
-        {label}
-      </label>
-      <input
-        className={inputStyle}
-        type={type}
-        name={name}
-        value={value}
-        id={id}
-        onChange={onChange}
-      />
-    </div>
-  );
-}
-
-InputText.propTypes = {
-  label: PropTypes.string,
-  type: PropTypes.string,
-  id: PropTypes.string,
-  name: PropTypes.string,
-  value: PropTypes.string,
-  onChange: PropTypes.func,
-  divStyle: PropTypes.string,
-  labelStyle: PropTypes.string,
-  inputStyle: PropTypes.string,
-};
-
-InputText.defaultProps = {
-  label: '',
-  type: '',
-  id: '',
-  name: '',
-  value: '',
-  onChange: null,
-  divStyle: '',
-  labelStyle: '',
-  inputStyle: '',
-};
+import InputText from '../InputText';
 
 class LoginPage extends Component {
   constructor() {
@@ -62,9 +19,9 @@ class LoginPage extends Component {
     this.closePopup = this.closePopup.bind(this);
   }
 
-  closePopup(newLink) {
+  closePopup() {
     this.setState({
-      linkTo: newLink,
+      linkTo: '',
       isPopup: false,
     });
   }
